CiviUnitTestCase - Extract Civi\Test\Api3DocTrait
This makes `CiviUnitTestCase` more readable, and it allows other tests to
benefit from these helpers (even if they don't use everything from
`CiviUnitTestCase`).
